cleanup: move workload evict to a subpackage
