+++ pkg/workload/admissionchecks.go
@@ -104,29 +104,6 @@ func SyncAdmittedCondition(w *kueue.Workload, now time.Time) bool {
 	return apimeta.SetStatusCondition(&w.Status.Conditions, newCondition)
 }
 
-// resetChecksOnEviction sets all AdmissionChecks to Pending
-func resetChecksOnEviction(w *kueue.Workload, now time.Time) {
-	checks := w.Status.AdmissionChecks
-	for i := range checks {
-		if checks[i].State == kueue.CheckStatePending {
-			continue
-		}
-		var retryCount *int32
-		if checks[i].State == kueue.CheckStateRetry {
-			tmpRetryCount := ptr.Deref(checks[i].RetryCount, 0) + 1
-			retryCount = new(tmpRetryCount)
-		}
-		checks[i] = kueue.AdmissionCheckState{
-			Name:                checks[i].Name,
-			State:               kueue.CheckStatePending,
-			Message:             "Reset to Pending after eviction. Previously: " + string(checks[i].State),
-			LastTransitionTime:  metav1.NewTime(now),
-			RequeueAfterSeconds: checks[i].RequeueAfterSeconds,
-			RetryCount:          retryCount,
-		}
-	}
-}
-
 // matchingChecks returns the list of admission checks in the given state.
 func matchingChecks(wl *kueue.Workload, s kueue.CheckState) []kueue.AdmissionCheckState {
 	matching := make([]kueue.AdmissionCheckState, 0, len(wl.Status.AdmissionChecks))
