+++ pkg/workload/evict/evict_test.go
@@ -1,0 +1,133 @@
+/*
+Copyright The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package evict
+
+import (
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	kueue "sigs.k8s.io/kueue/apis/kueue/v1beta2"
+	utiltestingapi "sigs.k8s.io/kueue/pkg/util/testing/v1beta2"
+)
+
+func TestIsEvictedByDeactivation(t *testing.T) {
+	cases := map[string]struct {
+		workload *kueue.Workload
+		want     bool
+	}{
+		"evicted condition doesn't exist": {
+			workload: utiltestingapi.MakeWorkload("test", "test").Obj(),
+		},
+		"evicted condition with false status": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadDeactivated,
+					Status: metav1.ConditionFalse,
+				}).
+				Obj(),
+		},
+		"evicted condition with PodsReadyTimeout reason": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
+					Status: metav1.ConditionTrue,
+				}).
+				Obj(),
+		},
+		"evicted condition with Deactivated reason": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadDeactivated,
+					Status: metav1.ConditionTrue,
+				}).
+				Obj(),
+			want: true,
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			got := IsEvictedByDeactivation(tc.workload)
+			if tc.want != got {
+				t.Errorf("Unexpected result from IsEvictedByDeactivation\nwant:%v\ngot:%v\n", tc.want, got)
+			}
+		})
+	}
+}
+
+func TestIsEvictedByPodsReadyTimeout(t *testing.T) {
+	cases := map[string]struct {
+		workload             *kueue.Workload
+		wantEvictedByTimeout bool
+		wantCondition        *metav1.Condition
+	}{
+		"evicted condition doesn't exist": {
+			workload: utiltestingapi.MakeWorkload("test", "test").Obj(),
+		},
+		"evicted condition with false status": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
+					Status: metav1.ConditionFalse,
+				}).
+				Obj(),
+		},
+		"evicted condition with Preempted reason": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadEvictedByPreemption,
+					Status: metav1.ConditionTrue,
+				}).
+				Obj(),
+		},
+		"evicted condition with PodsReadyTimeout reason": {
+			workload: utiltestingapi.MakeWorkload("test", "test").
+				Condition(metav1.Condition{
+					Type:   kueue.WorkloadEvicted,
+					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
+					Status: metav1.ConditionTrue,
+				}).
+				Obj(),
+			wantEvictedByTimeout: true,
+			wantCondition: &metav1.Condition{
+				Type:   kueue.WorkloadEvicted,
+				Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
+				Status: metav1.ConditionTrue,
+			},
+		},
+	}
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			gotCondition, gotEvictedByTimeout := IsEvictedByPodsReadyTimeout(tc.workload)
+			if tc.wantEvictedByTimeout != gotEvictedByTimeout {
+				t.Errorf("Unexpected evictedByTimeout from IsEvictedByPodsReadyTimeout\nwant:%v\ngot:%v\n",
+					tc.wantEvictedByTimeout, gotEvictedByTimeout)
+			}
+			if diff := cmp.Diff(tc.wantCondition, gotCondition,
+				cmpopts.IgnoreFields(metav1.Condition{}, "LastTransitionTime")); len(diff) != 0 {
+				t.Errorf("Unexpected condition from IsEvictedByPodsReadyTimeout: (-want,+got):\n%s", diff)
+			}
+		})
+	}
+}
