+++ pkg/workload/workload_test.go
@@ -1127,111 +1127,6 @@ func TestAssignmentClusterQueueState(t *testing.T) {
 	}
 }
 
-func TestIsEvictedByDeactivation(t *testing.T) {
-	cases := map[string]struct {
-		workload *kueue.Workload
-		want     bool
-	}{
-		"evicted condition doesn't exist": {
-			workload: utiltestingapi.MakeWorkload("test", "test").Obj(),
-		},
-		"evicted condition with false status": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadDeactivated,
-					Status: metav1.ConditionFalse,
-				}).
-				Obj(),
-		},
-		"evicted condition with PodsReadyTimeout reason": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
-					Status: metav1.ConditionTrue,
-				}).
-				Obj(),
-		},
-		"evicted condition with Deactivated reason": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadDeactivated,
-					Status: metav1.ConditionTrue,
-				}).
-				Obj(),
-			want: true,
-		},
-	}
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			got := IsEvictedByDeactivation(tc.workload)
-			if tc.want != got {
-				t.Errorf("Unexpected result from IsEvictedByDeactivation\nwant:%v\ngot:%v\n", tc.want, got)
-			}
-		})
-	}
-}
-
-func TestIsEvictedByPodsReadyTimeout(t *testing.T) {
-	cases := map[string]struct {
-		workload             *kueue.Workload
-		wantEvictedByTimeout bool
-		wantCondition        *metav1.Condition
-	}{
-		"evicted condition doesn't exist": {
-			workload: utiltestingapi.MakeWorkload("test", "test").Obj(),
-		},
-		"evicted condition with false status": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
-					Status: metav1.ConditionFalse,
-				}).
-				Obj(),
-		},
-		"evicted condition with Preempted reason": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadEvictedByPreemption,
-					Status: metav1.ConditionTrue,
-				}).
-				Obj(),
-		},
-		"evicted condition with PodsReadyTimeout reason": {
-			workload: utiltestingapi.MakeWorkload("test", "test").
-				Condition(metav1.Condition{
-					Type:   kueue.WorkloadEvicted,
-					Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
-					Status: metav1.ConditionTrue,
-				}).
-				Obj(),
-			wantEvictedByTimeout: true,
-			wantCondition: &metav1.Condition{
-				Type:   kueue.WorkloadEvicted,
-				Reason: kueue.WorkloadEvictedByPodsReadyTimeout,
-				Status: metav1.ConditionTrue,
-			},
-		},
-	}
-	for name, tc := range cases {
-		t.Run(name, func(t *testing.T) {
-			gotCondition, gotEvictedByTimeout := IsEvictedByPodsReadyTimeout(tc.workload)
-			if tc.wantEvictedByTimeout != gotEvictedByTimeout {
-				t.Errorf("Unexpected evictedByTimeout from IsEvictedByPodsReadyTimeout\nwant:%v\ngot:%v\n",
-					tc.wantEvictedByTimeout, gotEvictedByTimeout)
-			}
-			if diff := cmp.Diff(tc.wantCondition, gotCondition,
-				cmpopts.IgnoreFields(metav1.Condition{}, "LastTransitionTime")); len(diff) != 0 {
-				t.Errorf("Unexpected condition from IsEvictedByPodsReadyTimeout: (-want,+got):\n%s", diff)
-			}
-		})
-	}
-}
-
 func TestFlavorResourceUsage(t *testing.T) {
 	cases := map[string]struct {
 		info *Info
